% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_map.R
\name{as_panel_map}
\alias{as_panel_map}
\title{Create a panel_map from a data frame like object}
\usage{
as_panel_map(x, code_in, code_out, split_in)
}
\arguments{
\item{x}{Data frame}

\item{code_in}{Variable in \code{x} containing the source classification codes to convert from.}

\item{code_out}{Variable in \code{x} containing the destination classification codes to convert to.}

\item{split_in}{Variable in \code{x} containing the transformation weights to apply
to source data for each source code.}
}
\description{
Create a panel_map from a data frame like object
}
\examples{
panel_map
}
