% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-dict-tools.R
\name{check_cd_coverage}
\alias{check_cd_coverage}
\title{Check code-dict coverage}
\usage{
check_cd_coverage(data, code_dict, code_from)
}
\arguments{
\item{data}{<df> with data to be converted. Must contain \code{code_from} column.}

\item{code_dict}{<dataframe> with instructions for \code{code_from} into target schema.}

\item{code_from}{col containing codes to be converted}
}
\value{

}
\description{
Checks if every \code{code_from} in \code{data} has
at least one destination (row) in \code{code_dict}.
Warns if any \code{data$code_from} is not present in \code{code_dict$code_from}
}
\concept{code_dict helpers}
