% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformr-package.R
\docType{package}
\name{conformr-package}
\alias{conformr}
\alias{conformr-package}
\title{conformr: Tools for conforming data across classifications and statistics}
\description{
conformr provides tools for converting data between classification standards, and correcting discrepancies in reported and calculated statistics.
}
\author{
\strong{Maintainer}: Cynthia Huang \email{29718979+cynthiahqy@users.noreply.github.com} (\href{https://orcid.org/0000-0002-9218-987X}{ORCID})

}
\keyword{internal}
