% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Convert data between standards}
\usage{
convert(
  data,
  code_dict,
  code_from,
  code_to,
  values_from,
  weight_col,
  values_to = NULL
)
}
\arguments{
\item{data}{A data frame to convert. Can be grouped data.}

\item{code_dict}{A data frame with code correspondence between standards.}

\item{code_from}{A string specifying the shared column name in \code{data} and \code{code_dict}
with the codes to convert from.}

\item{code_to}{A string specifying which column in \code{code_dict} contains the
destination codes to convert to.}

\item{values_from}{A string specifying which column in \code{data} to get cell values from.}

\item{weight_col}{A string specifying which column in \code{code_dict} to get the
weights to distribute \code{values_from} between \code{code_from} and \code{code_to}}

\item{values_to}{A string specifying the name of the new column created to store
the distributed values. Defaults to "value_new"}
}
\value{

}
\description{
\code{convert()} maps numeric data from one category standard into another.
It maintains the total "amount" of data across one-to-one, many-to-one,
and one-to-many correspondences.

Learn more in \code{vignette("convert")}
}
\examples{


}
