% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-map-tools.R
\name{dm_check_values}
\alias{dm_check_values}
\title{Verify data map values}
\usage{
dm_check_values(data_map, code_in, code_out, weights, values_from)
}
\arguments{
\item{values_from}{}
}
\value{

}
\description{
Check that every \code{code_in}-\code{code_out} correspondence has appropriate \code{values_from}
to apply \code{weights} to.
In particular, check that one-to-many splits have \code{values_from} duplicated across
all \code{code_in}-\code{code_out}-\code{weight} rows within a particular \code{code_in}
}
\seealso{
\code{\link[=dm_check_codes]{dm_check_codes()}}, \code{\link[=dm_check_weights]{dm_check_weights()}}

Other data map tools: 
\code{\link{dm_check_codes}()},
\code{\link{dm_check_weights}()}
}
\concept{data map tools}
