% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-map-tools.R
\name{dm_check_weights}
\alias{dm_check_weights}
\title{Verify data/code map weights}
\usage{
dm_check_weights(
  map,
  code_in,
  code_out,
  weights,
  success_fun = NULL,
  error_fun = NULL
)
}
\arguments{
\item{map}{data map with correspondence between code_in and code_out}

\item{error_fun}{default assertr::error_report}

\item{succes_fun}{default assertr::success_continue}
}
\value{
map
}
\description{
Check that weights applied to given code_in/value_from add up to 1.
Ensures that no value is created or lost in the transformation process.
Each \code{value_from} numeric will be split into parts totalling 100\%
}
\details{
NA values of weights will (cause error/be fixed?)
}
\seealso{
\code{\link[=dm_check_values]{dm_check_values()}}, \code{\link[=dm_check_codes]{dm_check_codes()}}

Other data map tools: 
\code{\link{dm_check_codes}()},
\code{\link{dm_check_values}()}
}
\concept{data map tools}
