% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-dict-tools.R
\name{make_cd_equal}
\alias{make_cd_equal}
\title{Create code map from code correspondence with equal weight splits}
\usage{
make_cd_equal(codes, code_from, code_to, name_weight_col = NULL)
}
\arguments{
\item{codes}{df containing correspondence btween input and output codes}

\item{code_from}{column name of input code}

\item{code_to}{column name of output}
}
\value{

}
\description{
Generate code map for transforming values between classification codes,
using all distinct correspondence between two codes.
}
