% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_map-checkers.R
\name{new_panel_map}
\alias{new_panel_map}
\title{Low level constuctor for tbl_pm}
\usage{
new_panel_map(x, code_in, code_out, split_in, by = NA)
}
\arguments{
\item{x}{Data frame containing the following variables}

\item{code_in}{Variable that specifies the source classification categories.}

\item{code_out}{Variable that specifies the target classification categories.}

\item{split_in}{Variable that specifies the weights to apply to source values.}

\item{by}{Variables(s) that specify row grouping for "sub" panel maps.
For example, \verb{code_in = "ISICv4", by = "country"} retains \code{x$country}
and \code{x$ISICv4} in the panel_map.}
}
\value{
A \code{panel_map} object
}
\description{
Low level constuctor for tbl_pm
}
\author{
Cynthia Huang
}
