% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_map.R
\name{panel_map}
\alias{panel_map}
\title{Create a panel_map object}
\usage{
panel_map(x, code_in, code_out, split_in)
}
\arguments{
\item{code_in}{Variable in \code{x} containing the source classification codes to convert from.}

\item{code_out}{Variable in \code{x} containing the destination classification codes to convert to.}

\item{split_in}{Variable in \code{x} containing the transformation weights to apply
to source data for each source code.}

\item{key}{Variable(s)}
}
\value{

}
\description{
Create a panel_map object
}
\examples{
panel_map
}
