% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_NA.R
\name{sum_greedy}
\alias{sum_greedy}
\title{NA-preserving Sum of Vector Elements}
\usage{
sum_greedy(x)
}
\arguments{
\item{x}{vector}
}
\value{
\code{NA} if all elements in \code{x} are \code{NA}.
Otherwise, sum of all non-missing elements in vector.
}
\description{
Logical switching of \code{na.rm} argument for \code{\link[base]{sum}}
based on whether all inputs are \code{NA}.
}
\details{
Preserves returning of \code{NA} if all elements are \code{NA}.
Otherwise, behaves like \code{sum(x, na.rm=TRUE)} and
returns sum of non-missing elements.
}
\examples{
sum_NA(1:4)          ## returns 10
sum_NA(c(NA,NA,NA))  ## returns NA
sum_NA(c(2, NA, 32)) ## returns 34

}
