% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_panel_map.R
\name{transform_data_by_map}
\alias{transform_data_by_map}
\title{Check and apply panel_map transformation to data}
\usage{
transform_data_by_map(
  map,
  data,
  values_from,
  code_from,
  .validate_coverage = TRUE
)
}
\arguments{
\item{code_from}{}
}
\description{
Check and apply panel_map transformation to data
}
