% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_panel_map.R
\name{use_panel_map}
\alias{use_panel_map}
\title{Apply panel_map to data without checks}
\usage{
use_panel_map(map, data, values_from, code_from)
}
\arguments{
\item{map}{A data frame that contains a valid panel map for the transformation.}

\item{data}{A data frame for transformation.}

\item{values_from}{Variables in \code{data} containing the values to be converted.}

\item{code_from}{A character string of the variable containing the codes to convert from.
If the variables containing the source codes differ in \code{map} and \code{data}, use a named vector.
For example, \code{from_code = c("a" = "b")} will match codes in \code{map$a} to \code{data$b}.}

\item{to_code}{Variable in \code{map} containing the codes to convert to.}

\item{weights}{Variable in \code{map} containing the weights to split values by.}
}
\value{
The output has the following properties:
\itemize{
\item Groups are taken from \code{data_from}
}
}
\description{
A wrapper around a \code{{dplyr}} pipeline that takes a panel_map,
joins it with data, and transforms selected variables in that data according to
instructions in the panel map. Any groups in \code{data} are preserved.
}
