% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_map.R
\name{validate_panel_map}
\alias{validate_panel_map}
\title{A Validator for \code{panel_map} objects}
\usage{
validate_panel_map(x)
}
\arguments{
\item{pm}{}
}
\value{

}
\description{
A Validator for \code{panel_map} objects
}
