% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_panel_map.R
\name{check_cd_coverage}
\alias{check_cd_coverage}
\title{Check code-dict coverage}
\usage{
check_cd_coverage(map, data, code_in)
}
\arguments{
\item{map}{<panel_map>}

\item{data}{<df> with data to be converted. Must contain \code{code_in} column.}

\item{code_in}{A named character vector specifying the columns in \code{data} and \code{map}}
}
\value{

}
\description{
Checks if every \code{code_in} in \code{data} has
at least one destination (row) in \code{panel_map}.
Warns if any \code{data$code_in} is not present in \code{panel_map$code_in}
}
\concept{code_dict helpers}
