% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformr-package.R
\docType{package}
\name{conformr-package}
\alias{conformr}
\alias{conformr-package}
\title{conformr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Cynthia Huang \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
