% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive-convert.R
\name{convert}
\alias{convert}
\title{Convert data between standards}
\usage{
convert(
  data_map,
  code_in,
  code_out,
  weights,
  values_from,
  names_suffix = "_out",
  .check = TRUE,
  .sum_na = TRUE
)
}
\arguments{
\item{data_map}{A data frame with data to be converted,
and correspondence and weights between codes, for every data value}

\item{code_in}{Variable in \code{data_map} containing the codes to convert from}

\item{code_out}{Variable in \code{data_map} containing the destination codes to convert to.}

\item{weights}{Variable in \code{data_map} to get the
weights to distribute \code{values_from} between \code{code_in} and \code{code_out}}

\item{values_from}{\verb{<tidy-select>} columns in \code{data-map} to get cell values from.}

\item{names_suffix}{A string appended to every \code{value_from} column name to create column names for transformed values.
Defaults to "_out"}
}
\value{

}
\description{
\code{convert()} maps numeric data from one category standard into another,
and verifies that total "amount" of data across one-to-one, many-to-one,
and one-to-many correspondences.

Learn more in \code{vignette("convert")}
}
\examples{


}
