% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_panel_map.R
\name{use_panel_map}
\alias{use_panel_map}
\title{Use a panel map to transform data}
\usage{
use_panel_map(map, data, values_from, from_code, to_code, weights)
}
\arguments{
\item{map}{A data frame that contains a valid panel map for the transformation.}

\item{data}{A data frame for transformation.}

\item{values_from}{Variables in \code{data} containing the values to be converted.}

\item{from_code}{A character vector of the variable containing the codes to convert from.
If the variables containing the source codes differ in \code{map} and \code{data}, use a named vector.
For example, \code{from_code = c("a" = "b")} will match codes in \code{map$a} to \code{data$b}.
Only the column name \code{"a"} is kept in the output.}

\item{to_code}{Variable in \code{map} containing the codes to convert to.}

\item{weights}{Variable in \code{map} containing the weights to split values by.}
}
\value{
The output has the following properties:
\itemize{
\item Groups are taken from \code{data_from}
}
}
\description{
A light wrapper around a \code{{dplyr}} pipeline that takes a valid panel map,
joins it with data, and transforms selected variables in that data according to
instructions in the panel map. Any groups in \code{data} are preserved.
}
