% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_panel_map.R
\name{use_panel_map}
\alias{use_panel_map}
\title{Apply panel_map to data without checks}
\usage{
use_panel_map(
  map,
  data,
  values_from,
  from_code,
  to_code,
  weights,
  .suffix = NULL
)
}
\arguments{
\item{map}{A data frame that contains a valid panel map for the transformation.}

\item{data}{A data frame for transformation.}

\item{values_from}{Variable in \code{data} containing the values to be converted.}

\item{from_code}{A character string of the variable containing the codes to convert from.
Must be the same in \code{map} and \code{data}.}

\item{to_code}{Variable in \code{map} containing the codes to convert to.}

\item{weights}{Variable in \code{map} containing the weights to split values by.}

\item{.suffix}{A string appended to \code{value_from} column name to create column names for transformed values.
Defaults to \code{to_code}}
}
\value{
The output has the following properties:
\itemize{
\item Groups are taken from \code{data_from}
}
}
\description{
A wrapper around a \code{{dplyr}} pipeline that takes a panel_map,
joins it with data, and transforms selected variables in that data according to
instructions in the panel map. Any groups in \code{data} are preserved.
}
